requirements.txt
streamlit>=1.20
tensorflow>=2.10
numpy
pillow
opencv-python-headless
scikit-learn